# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Questline Deep Agent API
  description: |
    AI agent API that analyzes behavioral data to project life scenarios and
    generate concrete daily micro-actions. Follows Stripe-like conventions
    for agent consumption.

    Resource IDs are prefixed by type: `scen_`, `qst_`, `act_`, `apprv_`,
    `evt_`, `msg_`, `wrkr_`. All mutating endpoints require an
    `Idempotency-Key` header.
  version: "2026-03-01"
  contact:
    name: Questline Team
servers:
  - url: http://localhost:8000
    description: Local development
security:
  - BearerAuth: []
paths:
...
  /v1/scenarios:
    get:
      operationId: listScenarios
      summary: List scenarios
      tags: [Scenarios]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/StartingAfter"
        - $ref: "#/components/parameters/Expand"
      responses:
        "200":
          description: A list of scenarios
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScenarioList"
              example:
                object: list
                url: /v1/scenarios
                has_more: false
                data:
                  - id: scen_01
                    object: scenario
                    created: 1709510400
                    livemode: false
                    metadata: {}
                    title: Career Acceleration
                    horizon: 5yr
                    likelihood: most_likely
                    summary: Focused upskilling leads to a senior role within 5 years.
                    tags: [career, education]
                    patterns: [pat_01, pat_03]
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/scenarios/{scenario_id}:
    get:
      operationId: getScenario
      summary: Retrieve a scenario
      tags: [Scenarios]
      parameters:
        - name: scenario_id
          in: path
          required: true
          schema:
            type: string
            pattern: ^scen_
          example: scen_01
        - $ref: "#/components/parameters/Expand"
      responses:
        "200":
          description: A scenario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------- Quests ----------
...
  /v1/actions:
    post:
      operationId: createActions
      summary: Generate actions for a scenario
      tags: [Actions]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scenario_id]
              properties:
                scenario_id:
                  type: string
                  pattern: ^scen_
            example:
              scenario_id: scen_01
      responses:
        "201":
          description: Actions generated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Action"
              example:
                - id: act_x1
                  object: action
                  created: 1709510600
                  livemode: false
                  metadata: {}
                  scenario: scen_01
                  title: Review monthly budget
                  description: Audit discretionary spending against savings target.
                  data_ref: transactions.jsonl#L42
                  pattern_id: pat_01
                  rationale: Spending trend conflicts with savings goal.
                  compound_summary: null
        "400":
          $ref: "#/components/responses/BadRequest"

  # ---------- Approvals ----------
...
  /v1/workers:
    get:
      operationId: listWorkers
      summary: List workers
      tags: [Workers]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/StartingAfter"
        - $ref: "#/components/parameters/Expand"
      responses:
        "200":
          description: A list of workers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerList"
              example:
                object: list
                url: /v1/workers
                has_more: false
                data:
                  - id: wrkr_kg
                    object: worker
                    created: 1709500000
                    livemode: false
                    metadata: {}
                    label: KgWorker
                    status: idle
                    queue_depth: 0
                    last_error: null
                    skills:
                      - pattern_recognition
                      - graph_query
  /v1/workers/{worker_id}:
    get:
      operationId: getWorker
      summary: Retrieve a worker
      tags: [Workers]
      parameters:
        - name: worker_id
          in: path
          required: true
          schema:
            type: string
            pattern: ^wrkr_
          example: wrkr_kg
        - $ref: "#/components/parameters/Expand"
      responses:
        "200":
          description: A worker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API key prefixed with `sk_test_` (test mode) or `sk_live_` (live mode).
        Authentication is optional for local development.
  parameters:
    Limit:
      name: limit
      in: query
      description: Number of items to return (1-100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    StartingAfter:
      name: starting_after
      in: query
      description: Cursor for pagination — ID of the last item from previous page
      schema:
        type: string
    Expand:
      name: expand[]
      in: query
      description: Related objects to expand inline
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Unique key to ensure idempotent mutations
      schema:
        type: string
...
  schemas:
...
    ResourceBase:
      type: object
      required: [id, object, created, livemode, metadata]
      properties:
        id:
          type: string
        object:
          type: string
        created:
          type: integer
          description: Unix timestamp
        livemode:
          type: boolean
        metadata:
          type: object
          additionalProperties:
            type: string

    # ---------- Resources ----------
...
    Scenario:
      allOf:
        - $ref: "#/components/schemas/ResourceBase"
        - type: object
          properties:
            object:
              const: scenario
            title:
              type: string
            horizon:
              type: string
              enum: [1yr, 5yr, 10yr]
            likelihood:
              type: string
              enum: [most_likely, possible, aspirational]
            summary:
              type: string
            tags:
              type: array
              items:
                type: string
            patterns:
              type: array
              items:
                type: string
              description: Pattern IDs referenced by this scenario
...
    Action:
      allOf:
        - $ref: "#/components/schemas/ResourceBase"
        - type: object
          properties:
            object:
              const: action
            scenario:
              type: string
            title:
              type: string
            description:
              type: string
            data_ref:
              type: string
              description: "Reference to source data record, e.g. transactions.jsonl#L42"
            pattern_id:
              type: string
            rationale:
              type: string
            compound_summary:
              type: string
              nullable: true
...
    Worker:
      allOf:
        - $ref: "#/components/schemas/ResourceBase"
        - type: object
          properties:
            object:
              const: worker
            label:
              type: string
            status:
              type: string
              enum: [idle, busy, error, offline]
            queue_depth:
              type: integer
            last_error:
              type: string
              nullable: true
            skills:
              description: Skill names (expanded when requested)
              type: array
              items:
                type: string
...
    ListObject:
      type: object
      required: [object, data, has_more, url]
      properties:
        object:
          type: string
          const: list
        data:
          type: array
          items: {}
        has_more:
          type: boolean
        url:
          type: string
    ScenarioList:
      allOf:
        - $ref: "#/components/schemas/ListObject"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Scenario"
...
    WorkerList:
      allOf:
        - $ref: "#/components/schemas/ListObject"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Worker"

    # ---------- Error ----------
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [type, code, message]
          properties:
            type:
              type: string
              enum:
                - invalid_request_error
                - authentication_error
                - not_found_error
                - rate_limit_error
                - api_error
            code:
              type: string
            message:
              type: string
            param:
              type: string
              nullable: true
            doc_url:
              type: string
              nullable: true
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error:
              type: invalid_request_error
              code: parameter_invalid
              message: "limit must be between 1 and 100."
              param: limit
              doc_url: https://api.porthon.ai/docs/errors
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error:
              type: not_found_error
              code: resource_missing
              message: The requested resource does not exist.
              param: null
              doc_url: https://api.porthon.ai/docs/errors